Moved application source files in their own src directory. Changed Makefile and ui|irc/util.hpp for the new directory tree.
