+++ src/channel_buffer.cpp
