+++ src/error_buffer.cpp
