+++ src/message_stencil.cpp
