+++ src/message_stencil.hpp
