+++ src/session_buffer.cpp
