+++ src/unique_connection.hpp
